Move test files into src/ and add an assertion for files not covered
