+++ test/fixtures/test-jest-app/should-not-count-for-coverage.js
@@ -1,0 +1,4 @@
+// @flow
+module.exports = {
+  thisFileShouldNotCountForCoverage: () => {},
+};
