+++ test/fixtures/test-jest-app/src/__tests__/fail-browser.browser.js
