+++ test/fixtures/test-jest-app/src/__tests__/pass-browser.browser.js
