+++ test/fixtures/test-jest-app/src/dynamic-import.js
