+++ test/fixtures/test-jest-app/src/foo.js
