+++ test/fixtures/test-jest-app/src/main.js
