+++ test/fixtures/test-jest-app/src/snapshot.js
