+++ test/fixtures/test-jest-babel/src/__tests__/babel.js
