+++ test/fixtures/test-jest-babel/src/plugin.js
